requirements.txt
django>=4.2
djangorestframework>=3.14
mysqlclient>=2.2
djangorestframework-simplejwt>=5.2